[ refactor ] separate business logic on card bookmark
